Support multiple RandomX variants
This replaces the RandomWOW support with a copy of xmrig's RandomX
support (from the upcoming `evo` branch) which supports multiple
configurations at runtime.

This replaces the "random_wow" algorithm with a "randomx" algorithm that
takes variants (like the cryptonight algorithms).  Variant 0 (the
default when specifying no variant) is Monero's RandomX; variants 17 and
18 (using xmrig's variant values) are the RandomWOW and RandomXL (Loki)
variants.